Unit Tests Analyze JSON
- Renamed File
- Created Unit Tests
*Note* - Unit Test: test_approximate_string_matching_no_match should return no match but is matching apple to grape. This test will continue to fail until algorithm is updated to be more accurate.
- Updated script to require being directly to execute example code
